Made drop statements compatible to deleting the table before

+++ app/api2/database/migrations/2017_04_16_011415_create_shows_table.php
@@ -63,7 +63,10 @@ class CreateShowsTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('shows');
+		if(Schema::hasTable('shows'))
+		{
+			Schema::drop('shows');
+		}
 	}
 
 }
