+++ app/api2/database/migrations/2017_04_17_011415_create_playsheets_table.php
@@ -49,6 +49,9 @@ class CreatePlaysheetsTable extends Migration
      */
     public function down()
     {
-        Schema::drop('playsheets');
+		if(Schema::hasTable('playsheets'))
+		{
+			Schema::drop('playsheets');
+		}
     }
 }
