+++ app/api2/database/migrations/2017_04_18_011415_create_playitems_table.php
@@ -53,7 +53,10 @@ class CreatePlayitemsTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('playitems');
+		if(Schema::hasTable('playitems'))
+		{
+			Schema::drop('playitems');
+		}
 	}
 
 }
