+++ app/api2/database/migrations/2017_04_19_011410_create_membership_table.php
@@ -59,6 +59,9 @@ class CreateMembershipTable extends Migration
      */
     public function down()
     {
-        Schema::drop('membership');
+		if(Schema::hasTable('membership'))
+		{
+		    Schema::drop('membership');
+		}
     }
 }
