+++ app/api2/database/migrations/2017_04_19_011411_create_membership_years_table.php
@@ -56,6 +56,9 @@ class CreateMembershipYearsTable extends Migration
      */
     public function down()
     {
-        Schema::drop('membership_years');
+		if(Schema::hasTable('membership_years'))
+		{
+			Schema::drop('membership_years');
+		}
     }
 }
