+++ app/api2/database/migrations/2017_04_19_011412_create_user_table.php
@@ -36,6 +36,9 @@ class CreateUserTable extends Migration
      */
     public function down()
     {
-        Schema::drop('user');
+		if(Schema::hasTable('user'))
+		{
+        	Schema::drop('user');
+		}
     }
 }
