+++ app/api2/database/migrations/2017_04_19_011415_create_adlog_table.php
@@ -39,7 +39,10 @@ class CreateAdlogTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('adlog');
+		if(Schema::hasTable('adlog'))
+		{
+			Schema::drop('adlog');
+		}
 	}
 
 }
