+++ app/api2/database/migrations/2017_04_19_011415_create_djland_options_table.php
@@ -30,7 +30,10 @@ class CreateDjlandOptionsTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('djland_options');
+		if(Schema::hasTable('djland_options'))
+		{
+			Schema::drop('djland_options');
+		}
 	}
 
 }
