+++ app/api2/database/migrations/2017_04_19_011415_create_friends_table.php
@@ -34,7 +34,10 @@ class CreateFriendsTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('friends');
+		if(Schema::hasTable('friends'))
+		{
+			Schema::drop('friends');
+		}
 	}
 
 }
