+++ app/api2/database/migrations/2017_04_19_011415_create_group_members_table.php
@@ -34,6 +34,9 @@ class CreateGroupMembersTable extends Migration
      */
     public function down()
     {
-        Schema::drop('group_members');
+		if(Schema::hasTable('group_members'))
+		{
+			Schema::drop('group_members');
+		}
     }
 }
