+++ app/api2/database/migrations/2017_04_19_011415_create_hosts_table.php
@@ -28,7 +28,10 @@ class CreateHostsTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('hosts');
+		if(Schema::hasTable('hosts'))
+		{
+			Schema::drop('hosts');
+		}
 	}
 
 }
