+++ app/api2/database/migrations/2017_04_19_011415_create_library_songs_table.php
@@ -44,6 +44,9 @@ class CreateLibrarySongsTable extends Migration
      */
     public function down()
     {
-        Schema::drop('library_songs');
+		if(Schema::hasTable('library_songs'))
+		{
+        	Schema::drop('library_songs');
+		}
     }
 }
