+++ app/api2/database/migrations/2017_04_19_011415_create_library_table.php
@@ -45,7 +45,10 @@ class CreateLibraryTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('library');
+		if(Schema::hasTable('library'))
+		{
+			Schema::drop('library');
+		}
 	}
 
 }
