+++ app/api2/database/migrations/2017_04_19_011415_create_log_table.php
@@ -31,6 +31,9 @@ class CreateLogTable extends Migration
      */
     public function down()
     {
-        Schema::drop('log');
+		if(Schema::hasTable('log'))
+		{
+			Schema::drop('log');
+		}
     }
 }
