+++ app/api2/database/migrations/2017_04_19_011415_create_login_status_table.php
@@ -26,7 +26,10 @@ class CreateLoginStatusTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('login_status');
+		if(Schema::hasTable('login_status'))
+		{
+			Schema::drop('login_status');
+		}
 	}
 
 }
