+++ app/api2/database/migrations/2017_04_19_011415_create_member_resources_table.php
@@ -32,7 +32,10 @@ class CreateMemberResourcesTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('member_resources');
+		if(Schema::hasTable('member_resources'))
+		{
+			Schema::drop('member_resources');
+		}
 	}
 
 }
