+++ app/api2/database/migrations/2017_04_19_011415_create_member_show_table.php
@@ -29,7 +29,10 @@ class CreateMemberShowTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('member_show');
+		if(Schema::hasTable('member_show'))
+		{
+			Schema::drop('member_show');
+		}
 	}
 
 }
