+++ app/api2/database/migrations/2017_04_19_011415_create_membership_status_table.php
@@ -30,7 +30,9 @@ class CreateMembershipStatusTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('membership_status');
+		if(Schema::hasTable('membership_status')){
+			Schema::drop('membership_status');
+		}
 	}
 
 }
