+++ app/api2/database/migrations/2017_04_19_011415_create_podcast_episodes_table.php
@@ -41,7 +41,10 @@ class CreatePodcastEpisodesTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('podcast_episodes');
+		if(Schema::hasTable('podcast_episodes'))
+		{
+			Schema::drop('podcast_episodes');
+		}
 	}
 
 }
