+++ app/api2/database/migrations/2017_04_19_011415_create_show_times_table.php
@@ -35,6 +35,9 @@ class CreateShowTimesTable extends Migration
      */
     public function down()
     {
-        Schema::drop('show_times');
+		if(Schema::hasTable('show_times'))
+        {
+			Schema::drop('show_times');
+		}
     }
 }
