+++ app/api2/database/migrations/2017_04_19_011415_create_socan_table.php
@@ -30,6 +30,9 @@ class CreateSocanTable extends Migration
      */
     public function down()
     {
-        Schema::drop('socan');
+		if(Schema::hasTable('socan'))
+        {
+			Schema::drop('socan');
+		}
     }
 }
