+++ app/api2/database/migrations/2017_04_19_011415_create_social_table.php
@@ -30,7 +30,10 @@ class CreateSocialTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('social');
+		if(Schema::hasTable('social'))
+		{
+			Schema::drop('social');
+		}
 	}
 
 }
