+++ app/api2/database/migrations/2017_04_19_011415_create_songs_table.php
@@ -30,7 +30,10 @@ class CreateSongsTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('songs');
+		if(Schema::hasTable('songs'))
+		{
+			Schema::drop('songs');
+		}
 	}
 
 }
