+++ app/api2/database/migrations/2017_04_19_011415_create_special_events_table.php
@@ -35,7 +35,11 @@ class CreateSpecialEventsTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('special_events');
+		if(Schema::hasTable('special_events'))
+		{
+			Schema::drop('special_events');
+		}
+
 	}
 
 }
