+++ app/api2/database/migrations/2017_04_19_011415_create_submission_songs_table.php
@@ -44,6 +44,9 @@ class CreateSubmissionSongsTable extends Migration
      */
     public function down()
     {
-        Schema::drop('submission_songs');
+		if(Schema::hasTable('submission_songs'))
+		{
+        	Schema::drop('submission_songs');
+		}
     }
 }
