+++ app/api2/database/migrations/2017_04_19_011415_create_submissions_archive_table.php
@@ -38,7 +38,10 @@ class CreateSubmissionsArchiveTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('submissions_archive');
+		if(Schema::hasTable('submissions_archive'))
+		{
+			Schema::drop('submissions_archive');
+		}
 	}
 
 }
