+++ app/api2/database/migrations/2017_04_19_011415_create_submissions_rejected_table.php
@@ -31,7 +31,10 @@ class CreateSubmissionsRejectedTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('submissions_rejected');
+		if(Schema::hasTable('submissions_rejected'))
+		{
+			Schema::drop('submissions_rejected');
+		}
 	}
 
 }
