+++ app/api2/database/migrations/2017_04_19_011415_create_submissions_table.php
@@ -58,7 +58,10 @@ class CreateSubmissionsTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('submissions');
+		if(Schema::hasTable('submissions'))
+		{
+			Schema::drop('submissions');
+		}
 	}
 
 }
