+++ app/api2/database/migrations/2017_04_19_011415_create_types_format_table.php
@@ -28,7 +28,10 @@ class CreateTypesFormatTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('types_format');
+		if(Schema::hasTable('types_format'))
+		{
+			Schema::drop('types_format');
+		}
 	}
 
 }
