+++ app/api2/database/migrations/2017_04_19_011415_create_uploads_table.php
@@ -36,7 +36,10 @@ class CreateUploadsTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('uploads');
+		if(Schema::hasTable('uploads'))
+		{
+			Schema::drop('uploads');
+		}
 	}
 
 }
