+++ app/api2/database/migrations/2017_04_19_011415_create_user_backup_table.php
@@ -13,7 +13,7 @@ class CreateUserBackupTable extends Migration {
 	public function up()
 	{
 		//TODO: can we delete this table? It seems to have been left around
-		//From a change in the past 
+		//From a change in the past
 		Schema::create('user_backup', function(Blueprint $table)
 		{
 			$table->integer('userid')->unsigned()->default(0);
@@ -36,7 +36,10 @@ class CreateUserBackupTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('user_backup');
+		if(Schema::hasTable('user_backup'))
+		{
+			Schema::drop('user_backup');
+		}
 	}
 
 }
