+++ app/api2/database/migrations/2017_04_19_011417_create_genres_table.php
@@ -33,6 +33,9 @@ class CreateGenresTable extends Migration
      */
     public function down()
     {
-        Schema::drop('genres');
+		if(Schema::hasTable('genres'))
+		{
+        	Schema::drop('genres');
+		}
     }
 }
