+++ app/api2/database/migrations/2017_04_19_011418_create_subgenres_table.php
@@ -33,6 +33,9 @@ class CreateSubgenresTable extends Migration
      */
     public function down()
     {
-        Schema::drop('subgenres');
+		if(Schema::hasTable('subgenres'))
+		{
+        	Schema::drop('subgenres');
+		}
     }
 }
