+++ app/api2/database/migrations/2017_05_07_202430_create_loglevel_table.php
@@ -32,6 +32,9 @@ class CreateLoglevelTable extends Migration
         Schema::table('log', function (Blueprint $table) {
             $table->dropForeign('fk_log_1');
         });
-        Schema::drop('loglevel');
+		if(Schema::hasTable('loglevel'))
+		{
+        	Schema::drop('loglevel');
+		}
     }
 }
