+++ app/api2/database/migrations/2017_05_10_090022_create_library_edits.php
@@ -52,6 +52,9 @@ class CreateLibraryEdits extends Migration
      */
     public function down()
     {
-        Schema::drop('library_edits');
+		if(Schema::hasTable('library_edits'))
+		{
+        	Schema::drop('library_edits');
+		}
     }
 }
